testsuite: 	tests for blocked user and some corner cases. dkimsign: 	temporary ctlfile is not removed when testing. filterlib: 	new function free_on_exit() to delay freeing filter-response. zdkimfilter: 	new function some_dwa_cleanup() to cleanup dwa separately. 	parse message header to collect data also if not signing. 	fixed check in block_user(). 	syslog line on altering block list raised to LOG_CRIT.
git-svn-id: http://www.tana.it/svn/zdkimfilter/trunk@67 da2204c7-d1c4-4af5-8207-86abab2c59ff

+++ src/dkimsign.c
@@ -949,7 +949,8 @@ int main(int argc, char *argv[])
 
 	if (ctlfile)
 	{
-		unlink(ctlfile);
+		if (!no_needless_logging)
+			unlink(ctlfile);
 		free(ctlfile);
 	}
